The data are recalculated
The bug is fixed in calculation of particle fluxes at the air-ground
interface. The data are recalculated using the new patch-02 of
Geant4.9.5. The relative changes in particle fluxes are less than 3% at
high altitudes. The particle fluxes are lower by about 5-10% at the
air-ground level. The geomagnetic cutoff effect for protons is found to
be lower and constitutes 0.57, 0.6 and 0.63 for low, average and high
levels of solar activity (instead of 0.61, 0.64 and 0.67). The changes
in other parameters are insignificant (atenuation lengths, angle
distributions).